milady: fix final blocking findings for PR 417
